+++ src/main/java/com/cadiducho/bot/modules/pole/cmds/admin/BanUserCMD.java
@@ -20,7 +20,7 @@ import java.util.Optional;
         aliases = "/banuser",
         hidden = true,
         arguments = {
-                @Argument(name = "usuario", type = Integer.class, description = "ID del usuario"),
+                @Argument(name = "usuario", type = Long.class, description = "ID del usuario"),
                 @Argument(name = "grupo", type = Long.class, description = "ID del grupo donde cometió la falta"),
                 @Argument(name = "razon", type = String.class, description = "La razón")
         })
